pyproject.toml
[tool.poetry]
name = "selenium-worker"
version = "0.1.0"
description = ""
authors = ["Irishey <markirill21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.3"
undetected-chromedriver = "^3.1.5"
python-dotenv = "^0.20.0"
Flask = "^2.1.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
